+++ packages/client-ledger/RELEASES.md
@@ -187,3 +187,13 @@ Based on:
 - [typescript v1.0.0-alpha.18] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.18] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.18 - packages/client-ledger
+
+## 2024-08-10 00:15:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.360.0 (2.392.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.19] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.19] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.19 - packages/client-ledger